# bff-server/swagger.yaml
openapi: 3.0.3
info:
  title: Moodle BFF API
  description: Backend For Frontend API for Moodle integration
  version: 1.0.0
  contact:
    name: Moodle BFF API Support
servers:
  - url: https://52.194.117.196
    description: Production server (HTTPS via nginx)
  - url: http://52.194.117.196
    description: Development server (HTTP, redirects to HTTPS)
  - url: http://localhost:3001
    description: Local development (direct access)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Cognito JWT認証を使用します。

        1. Cognitoでログインして IDToken を取得
        2. "Authorize" ボタンをクリック
        3. IDToken を入力（"Bearer" プレフィックスは不要）

        トークン取得方法:
        ```bash
        # AWS CLIで取得
        aws cognito-idp initiate-auth \
          --auth-flow USER_PASSWORD_AUTH \
          --client-id 23jacbr6nk4baiftjueddmr4kb \
          --auth-parameters USERNAME=your-email,PASSWORD=your-password \
          --region ap-northeast-1

        # またはスクリプトを使用
        ./get-token.sh your-email your-password
        node test-auth.js your-email your-password
        ```
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
...
paths:
...
  /api/admin/cognito-users:
    post:
      tags:
        - Admin
      summary: Create/Delete Cognito users (Admin only)
      description: |
        Bulk create or delete users in AWS Cognito. This endpoint is restricted to users with the "admin" Cognito group.

        **For creating users:**
        - A Cognito account will be created
        - A temporary password will be sent to the user's email
        - Optionally, the user can be added to a Cognito group

        **For deleting users:**
        - Set `deleteFlag: true` in the record
        - Only email or username is required for deletion

        **Required fields (create):**
        - email: User's email address
        - username: Unique username for the user

        **Required fields (delete):**
        - email or username: To identify the user to delete
        - deleteFlag: true

        **Optional fields:**
        - group: Cognito group name (e.g., "admin", "Teachers", "Students")
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - records
              properties:
                records:
                  type: array
                  items:
                    type: object
                    properties:
                      email:
                        type: string
                        format: email
                        example: "user@example.com"
                      username:
                        type: string
                        example: "testuser01"
                      group:
                        type: string
                        example: "Students"
                        description: "Optional Cognito group to add the user to"
                      deleteFlag:
                        type: boolean
                        example: false
                        description: "Set to true to delete the user instead of creating"
      responses:
        '200':
          description: Users created/deleted (may include partial failures)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: True if all operations were successful
                    example: true
                  recordsProcessed:
                    type: integer
                    description: Number of users successfully created or deleted
                    example: 8
                  recordsDeleted:
                    type: integer
                    description: Number of users successfully deleted
                    example: 2
                  recordsFailed:
                    type: integer
                    description: Number of users that failed to process
                    example: 0
                  message:
                    type: string
                    example: "6件のユーザーを作成、2件のユーザーを削除しました"
                  errors:
                    type: array
                    description: Details of any errors that occurred
                    items:
                      type: object
                      properties:
                        row:
                          type: integer
                          description: Row number in the input array (1-indexed)
                        message:
                          type: string
                          description: Error message
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - admin privileges required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "管理者権限が必要です。この操作は管理者のみ実行できます。"
        '500':
          description: Failed to create users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Admin
      summary: List all Cognito users with groups (Admin only)
      description: |
        Get a complete list of all users in the Cognito User Pool with their group memberships.
        This endpoint is restricted to users with the "admin" Cognito group.

        Features:
        - Fetches ALL users with pagination support (no 60-user limit)
        - Efficiently retrieves group memberships by querying groups first, then members
        - Avoids per-user API calls for better performance

        Note: For large user pools, this operation may take several seconds.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list with groups retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        username:
                          type: string
                          example: "testuser01"
                        email:
                          type: string
                          example: "user@example.com"
                        status:
                          type: string
                          description: "User status (CONFIRMED, FORCE_CHANGE_PASSWORD, etc.)"
                          example: "CONFIRMED"
                        enabled:
                          type: boolean
                          example: true
                        createdAt:
                          type: string
                          format: date-time
                          example: "2024-01-15T10:30:00Z"
                        groups:
                          type: array
                          description: "List of Cognito groups this user belongs to"
                          items:
                            type: string
                          example: ["Students", "premium-users"]
                  total:
                    type: integer
                    description: "Total number of users retrieved"
                    example: 150
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - admin privileges required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "管理者権限が必要です。この操作は管理者のみ実行できます。"
        '500':
          description: Failed to list users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/students:
    get:
      tags:
        - Admin
      summary: List students from Cognito 'student' group with last login information (Admin/Coach)
      description: |
        Get a list of students from Cognito 'student' group with their Moodle last login information.
        Admin: returns all students in 'student' group
        Coach: returns only assigned students from 'student' group

        **Note:** This endpoint now only returns users who are in the Cognito 'student' group,
        excluding users with 'admin' or 'coach' roles.

        Features:
        - Fetches users from Cognito 'student' group
        - Enriches with Moodle user information
        - Calculates formatted last access time (e.g., "5分前", "3時間前", "7日前")
        - Provides flags for inactive users and new users
        - Returns statistics summary

        Flags:
        - `inactive_over_month`: true if user hasn't logged in for over 30 days (or never logged in)
        - `new_user`: true if user account was created within the last 30 days

        Results are sorted by most recent access first.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Student list retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  students:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Moodle user ID
                          example: 3
                        username:
                          type: string
                          example: "student01"
                        email:
                          type: string
                          example: "student01@example.com"
                        firstname:
                          type: string
                          example: "太郎"
                        lastname:
                          type: string
                          example: "山田"
                        fullname:
                          type: string
                          example: "山田 太郎"
                        lastaccess:
                          type: integer
                          description: Last access timestamp (Unix seconds, 0 if never logged in)
                          example: 1735689600
                        lastaccess_formatted:
                          type: string
                          description: Human-readable last access time
                          example: "5分前"
                        firstaccess:
                          type: integer
                          description: First access timestamp (Unix seconds)
                          example: 1704067200
                        suspended:
                          type: boolean
                          description: Whether the user is suspended
                          example: false
                        auth:
                          type: string
                          description: Authentication method
                          example: "manual"
                        inactive_over_month:
                          type: boolean
                          description: True if user hasn't logged in for over 30 days or never logged in
                          example: false
                        new_user:
                          type: boolean
                          description: True if user account was created within the last 30 days
                          example: true
                  stats:
                    type: object
                    description: Summary statistics
                    properties:
                      total:
                        type: integer
                        description: Total number of students
                        example: 50
                      inactive_over_month:
                        type: integer
                        description: Number of students inactive for over 1 month
                        example: 12
                      new_users:
                        type: integer
                        description: Number of new users (created within last month)
                        example: 8
                      never_logged_in:
                        type: integer
                        description: Number of users who never logged in
                        example: 3
                  timestamp:
                    type: string
                    format: date-time
                    description: Timestamp when this data was retrieved
                    example: "2024-01-15T10:30:00Z"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - admin privileges required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "管理者権限が必要です。この操作は管理者のみ実行できます。"
        '500':
          description: Failed to fetch students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== COACHING ENDPOINTS ====================
...